back-end/auth_service/requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.7
daphne==4.1.2
Django==4.2.10
django-cors-headers==4.3.1
django-crontab==0.7.1
django-elasticsearch-dsl==8.0
django-extensions==3.2.3
django-phonenumber-field==7.3.0
django-sslserver==0.22
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
easy-colors==0.0.1
elastic-transport==8.15.0
elasticsearch==8.15.0
elasticsearch-dsl==8.15.1
frozenlist==1.4.1
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.6
importlib_metadata==7.1.0
incremental==22.10.0
Markdown==3.6
MarkupSafe==2.1.5
msgpack==1.0.8
multidict==6.0.5
packaging==24.1
phonenumberslite==8.13.36
pillow==10.2.0
postgres==4.0
psycopg2-binary==2.9.9
psycopg2-pool==1.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
pyotp==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
qrcode==7.3.1
redis==5.0.4
requests==2.31.0
secure-smtplib==0.1.1
send-mail==0.0.2
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.4
twilio==8.13.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.9.0
urllib3==2.2.0
Werkzeug==3.0.1
yarl==1.9.4
zipp==3.18.2
zope.interface==6.4.post2
logstash-formatter==0.5.17
python-logstash==0.4.8
shortuuid==1.0.11
python-json-logger==2.0.7
django-redis==5.4.0